+++ packages/components-js/tests/vrt/specs/states/checkbox.vrt.ts
@@ -28,25 +28,16 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
     <p-checkbox name="some-name">
       <span slot="label">
         Slotted label
-        <span>
-          and some slotted, deeply nested <a href="#">anchor</a>.
-        </span>
       </span>
     </p-checkbox>
     <p-checkbox label="State error" state="error" name="some-name">
       <span slot="message">
         Slotted error message
-        <span>
-          and some slotted, deeply nested <a href="#">anchor</a>.
-        </span>
       </span>
     </p-checkbox>
     <p-checkbox label="State success" state="success" name="some-name">
       <span slot="message">
         Slotted success message
-        <span>
-          and some slotted, deeply nested <a href="#">anchor</a>.
-        </span>
       </span>
     </p-checkbox>`;